Added players list to team page and moved services
