+++ CS179-Serverside/SelectSports/src/app/app.module.ts
@@ -12,7 +12,7 @@ import {HttpClientModule} from '@angular/common/http';
 import { HomepageComponent } from './homepage/homepage.component';
 import { AuthService } from './Services/auth.service';
 import { UserProfileEditComponent } from './user-profile-edit/user-profile-edit.component';
-import { FilterLeaguePipe } from './filter-league.pipe';
+import { FilterLeaguePipe } from './Services/filter-league.pipe';
 import { SearchBarComponent } from './search-bar/search-bar.component';
 import { SearchResultComponent } from './search-result/search-result.component';
 import { TeamDetailComponent } from './team-detail/team-detail.component';
